Modularized backend routes.
New routes:
- authRoutes.js
- userRoutes.js
- socialRoutes.js
-- postRoutes.js, commentRoutes.js, and likeRoutes.js all added here to clean things up on the backend
